third_party/WebKit/LayoutTests/http/tests/inspector/web-socket-frame-error-expected.txt

CONSOLE ERROR: line 8: WebSocket connection to 'ws://localhost:8000/websocket/tests/hybi/does_not_exist' failed: Error during WebSocket handshake: Unexpected response code: 404
Closed.
Tests that WebSocketFrames errors are visible to Web Inspector.

1: Error during WebSocket handshake: Unexpected response code: 404
